redis: Rename `messages` module to `positions`
